added binary inclusion for the SWI-Prolog home

+++ Sources/LogicKit/LogicKit.c++
@@ -1,9 +1,9 @@
-#include <iterator>
 #include <sys/syslog.h>
 #include <syslog.h>
 
 #include <assert.h>
 
+#include <mutex>
 #include <algorithm>
 #include <vector>
 #include <string>
@@ -13,6 +13,7 @@
 #include <SWI-Prolog.h>
 #include <SWI-cpp2.h>
 
+#include <SWIPrologHome.h>
 #include "LogicKit.h++"
 
 namespace looe::LegalXML::LogicKit
@@ -33,7 +34,7 @@ PrologVM::PrologVM(const std::string &argv0)
     : args({ argv0, "--home=/usr/lib64/swipl-9.2.6" }),
       cArgs(map(
         this->args, +[](std::string &i) -> char * { return &i[0]; })),
-      engine(this->cArgs.size(), this->cArgs.data())
+      engine(boost::numeric_cast<int>(this->cArgs.size()), this->cArgs.data())
 {
   assert(PL_is_initialised(nullptr, nullptr));
 }
